shore(overview): rename data fetching hook and update imports
